requirements.txt
Flask==2.2.2
Flask-WTF==0.15.1
Flask-PyMongo==2.3.0
WTForms==2.3.3
bootstrap-flask==1.6.0
Werkzeug==2.2.3
email-validator==1.1.3
requests
python-dotenv